+++ pyvisim/__init__.py
@@ -3,8 +3,8 @@ pyvisim: A Python library for image similarity analysis using Image Encoders and
 """
 
 __all__ = [
-    "clustering",
     "datasets",
+    "distance",
     "encoders",
     "features",
     "eval",