+++ pyvisim/neural_networks/siamese/__init__.py
@@ -1,11 +1,9 @@
 from .backbones import ResNetBackbone
 from .contrastive_siamese_network import ContrastiveSiameseNetwork
 from .pairwise_siamese_network import PairwiseSiameseNetwork
-from .siamese_neural_network import SiameseNeuralNetwork
 
 __all__ = [
     "ContrastiveSiameseNetwork",
     "PairwiseSiameseNetwork",
     "ResNetBackbone",
-    "SiameseNeuralNetwork",
 ]